feat-#10: move activity repository to db module
