+++ src/catalog/activity/commands/handlers/create-activity.handler.ts
@@ -3,7 +3,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 
 import { RpcExceptionService } from '../../../../utils/exception-handling';
 import { ErrorValidationService } from '../../../../utils/error-validation';
-import { ActivityRepository } from '../../repositories';
+import { ActivityRepository } from '../../../../db/repositories';
 import { ActivityEntity } from '../../../../db/entities/activity.entity';
 import { CreateActivityCommand } from '../impl';
 import { EsService } from '../../../../es/es.service';
