+++ src/db/repositories/activity.repository.ts
@@ -1,6 +1,6 @@
 import { EntityRepository, Repository } from 'typeorm';
 
-import { ActivityEntity } from '../../../db/entities/activity.entity';
+import { ActivityEntity } from '../entities/activity.entity';
 
 @EntityRepository(ActivityEntity)
 export class ActivityRepository extends Repository<ActivityEntity> {}
